+++ src/terminal.go
@@ -97,6 +97,7 @@ type itemLine struct {
 	label    string
 	queryLen int
 	width    int
+	bar      bool
 	result   Result
 }
 
@@ -161,6 +162,7 @@ type Terminal struct {
 	header             []string
 	header0            []string
 	ellipsis           string
+	scrollbar          string
 	ansi               bool
 	tabstop            int
 	margin             [4]sizeSpec
@@ -311,6 +313,7 @@ const (
 	actToggleSort
 	actTogglePreview
 	actTogglePreviewWrap
+	actTransformPrompt
 	actTransformQuery
 	actPreview
 	actChangePreview
@@ -631,6 +634,15 @@ func NewTerminal(opts *Options, eventBox *util.EventBox) *Terminal {
 		}
 		t.separator, t.separatorLen = t.ansiLabelPrinter(bar, &tui.ColSeparator, true)
 	}
+	if opts.Scrollbar == nil {
+		if t.unicode {
+			t.scrollbar = "│"
+		} else {
+			t.scrollbar = "|"
+		}
+	} else {
+		t.scrollbar = *opts.Scrollbar
+	}
 
 	_, t.hasLoadActions = t.keymap[tui.Load.AsEvent()]
 
@@ -760,6 +772,27 @@ func (t *Terminal) parsePrompt(prompt string) (func(), int) {
 
 func (t *Terminal) noInfoLine() bool {
 	return t.infoStyle != infoDefault
+}
+
+func (t *Terminal) getScrollbar() (int, int) {
+	total := t.merger.Length()
+	if total == 0 {
+		return 0, 0
+	}
+
+	maxItems := t.maxItems()
+	barLength := util.Max(1, maxItems*maxItems/total)
+	if total <= maxItems {
+		return 0, 0
+	}
+
+	var barStart int
+	if total == maxItems {
+		barStart = 0
+	} else {
+		barStart = (maxItems - barLength) * t.offset / (total - maxItems)
+	}
+	return barLength, barStart
 }
 
 // Input returns current query string
@@ -1348,6 +1381,7 @@ func (t *Terminal) printHeader() {
 
 func (t *Terminal) printList() {
 	t.constrain()
+	barLength, barStart := t.getScrollbar()
 
 	maxy := t.maxItems()
 	count := t.merger.Length() - t.offset
@@ -1361,7 +1395,7 @@ func (t *Terminal) printList() {
 			line--
 		}
 		if i < count {
-			t.printItem(t.merger.Get(i+t.offset), line, i, i == t.cy-t.offset)
+			t.printItem(t.merger.Get(i+t.offset), line, i, i == t.cy-t.offset, i >= barStart && i < barStart+barLength)
 		} else if t.prevLines[i] != emptyLine {
 			t.prevLines[i] = emptyLine
 			t.move(line, 0, true)
@@ -1369,7 +1403,7 @@ func (t *Terminal) printList() {
 	}
 }
 
-func (t *Terminal) printItem(result Result, line int, i int, current bool) {
+func (t *Terminal) printItem(result Result, line int, i int, current bool, bar bool) {
 	item := result.item
 	_, selected := t.selected[item.Index()]
 	label := ""
@@ -1385,13 +1419,24 @@ func (t *Terminal) printItem(result Result, line int, i int, current bool) {
 
 	// Avoid unnecessary redraw
 	newLine := itemLine{current: current, selected: selected, label: label,
-		result: result, queryLen: len(t.input), width: 0}
+		result: result, queryLen: len(t.input), width: 0, bar: bar}
 	prevLine := t.prevLines[i]
+	printBar := func() {
+		if len(t.scrollbar) > 0 && bar != prevLine.bar {
+			t.prevLines[i].bar = bar
+			t.move(line, t.window.Width()-1, true)
+			if bar {
+				t.window.CPrint(tui.ColScrollbar, t.scrollbar)
+			}
+		}
+	}
+
 	if prevLine.current == newLine.current &&
 		prevLine.selected == newLine.selected &&
 		prevLine.label == newLine.label &&
 		prevLine.queryLen == newLine.queryLen &&
 		prevLine.result == newLine.result {
+		printBar()
 		return
 	}
 
@@ -1425,6 +1470,7 @@ func (t *Terminal) printItem(result Result, line int, i int, current bool) {
 	if fillSpaces > 0 {
 		t.window.Print(strings.Repeat(" ", fillSpaces))
 	}
+	printBar()
 	t.prevLines[i] = newLine
 }
 
@@ -2034,8 +2080,10 @@ func (t *Terminal) redraw(clear bool) {
 
 func (t *Terminal) executeCommand(template string, forcePlus bool, background bool, captureFirstLine bool) string {
 	line := ""
-	valid, list := t.buildPlusList(template, forcePlus)
-	if !valid {
+	valid, list := t.buildPlusList(template, forcePlus, false)
+	// captureFirstLine is used for transform-{prompt,query} and we don't want to
+	// return an empty string in those cases
+	if !valid && !captureFirstLine {
 		return line
 	}
 	command := t.replacePlaceholder(template, forcePlus, string(t.input), list)
@@ -2093,10 +2141,10 @@ func (t *Terminal) currentItem() *Item {
 	return nil
 }
 
-func (t *Terminal) buildPlusList(template string, forcePlus bool) (bool, []*Item) {
+func (t *Terminal) buildPlusList(template string, forcePlus bool, forceEvaluation bool) (bool, []*Item) {
 	current := t.currentItem()
 	slot, plus, query := hasPreviewFlags(template)
-	if !(!slot || query && len(t.input) > 0 || (forcePlus || plus) && len(t.selected) > 0) {
+	if !forceEvaluation && !(!slot || query || (forcePlus || plus) && len(t.selected) > 0) {
 		return current != nil, []*Item{current, current}
 	}
 
@@ -2421,7 +2469,7 @@ func (t *Terminal) Loop() {
 
 	refreshPreview := func(command string) {
 		if len(command) > 0 && t.isPreviewEnabled() {
-			_, list := t.buildPlusList(command, false)
+			_, list := t.buildPlusList(command, false, false)
 			t.cancelPreview()
 			t.previewBox.Set(reqPreviewEnqueue, previewRequest{command, t.pwindow, t.evaluateScrollOffset(), list})
 		}
@@ -2540,6 +2588,7 @@ func (t *Terminal) Loop() {
 			t.eventChan <- t.tui.GetChar()
 		}
 	}()
+	dragging := false
 	for looping {
 		var newCommand *string
 		var reloadSync bool
@@ -2649,7 +2698,7 @@ func (t *Terminal) Loop() {
 				if t.hasPreviewer() {
 					togglePreview(!t.previewer.enabled)
 					if t.previewer.enabled {
-						valid, list := t.buildPlusList(t.previewOpts.command, false)
+						valid, list := t.buildPlusList(t.previewOpts.command, false, false)
 						if valid {
 							t.cancelPreview()
 							t.previewBox.Set(reqPreviewEnqueue,
@@ -2664,6 +2713,10 @@ func (t *Terminal) Loop() {
 					t.previewed.version = 0
 					req(reqPreviewRefresh)
 				}
+			case actTransformPrompt:
+				prompt := t.executeCommand(a.a, false, true, true)
+				t.prompt, t.promptLen = t.parsePrompt(prompt)
+				req(reqPrompt)
 			case actTransformQuery:
 				query := t.executeCommand(a.a, false, true, true)
 				t.input = []rune(query)
@@ -2993,7 +3046,6 @@ func (t *Terminal) Loop() {
 				} else if t.window.Enclose(my, mx) {
 					mx -= t.window.Left()
 					my -= t.window.Top()
-					mx = util.Constrain(mx-t.promptLen, 0, len(t.input))
 					min := 2 + len(t.header)
 					if t.noInfoLine() {
 						min--
@@ -3009,6 +3061,7 @@ func (t *Terminal) Loop() {
 							my = h - my - 1
 						}
 					}
+					dragging = me.Down && (dragging || my >= min && mx == t.window.Width()-1)
 					if me.Double {
 						// Double-click
 						if my >= min {
@@ -3016,7 +3069,21 @@ func (t *Terminal) Loop() {
 								return doActions(actionsFor(tui.DoubleClick))
 							}
 						}
+					} else if dragging && my >= min {
+						barLength, barStart := t.getScrollbar()
+						if barLength > 0 {
+							maxItems := t.maxItems()
+							if newBarStart := util.Constrain(my-min-barLength/2, 0, maxItems-barLength); newBarStart != barStart {
+								total := t.merger.Length()
+								prevOffset := t.offset
+								// barStart = (maxItems - barLength) * t.offset / (total - maxItems)
+								t.offset = int(math.Ceil(float64(newBarStart) * float64(total-maxItems) / float64(maxItems-barLength)))
+								t.cy = t.offset + t.cy - prevOffset
+								req(reqList)
+							}
+						}
 					} else if me.Down {
+						mx = util.Constrain(mx-t.promptLen, 0, len(t.input))
 						if my == t.promptLine() && mx >= 0 {
 							// Prompt
 							t.cx = mx + t.xoffset
@@ -3036,7 +3103,7 @@ func (t *Terminal) Loop() {
 			case actReload, actReloadSync:
 				t.failed = nil
 
-				valid, list := t.buildPlusList(a.a, false)
+				valid, list := t.buildPlusList(a.a, false, false)
 				if !valid {
 					// We run the command even when there's no match
 					// 1. If the template doesn't have any slots