+++ src/tui/tui.go
@@ -269,6 +269,7 @@ type ColorTheme struct {
 	Selected     ColorAttr
 	Header       ColorAttr
 	Separator    ColorAttr
+	Scrollbar    ColorAttr
 	Border       ColorAttr
 	BorderLabel  ColorAttr
 	PreviewLabel ColorAttr
@@ -466,6 +467,7 @@ var (
 	ColInfo                 ColorPair
 	ColHeader               ColorPair
 	ColSeparator            ColorPair
+	ColScrollbar            ColorPair
 	ColBorder               ColorPair
 	ColPreview              ColorPair
 	ColPreviewBorder        ColorPair
@@ -489,7 +491,6 @@ func EmptyTheme() *ColorTheme {
 		Cursor:       ColorAttr{colUndefined, AttrUndefined},
 		Selected:     ColorAttr{colUndefined, AttrUndefined},
 		Header:       ColorAttr{colUndefined, AttrUndefined},
-		Separator:    ColorAttr{colUndefined, AttrUndefined},
 		Border:       ColorAttr{colUndefined, AttrUndefined},
 		BorderLabel:  ColorAttr{colUndefined, AttrUndefined},
 		Disabled:     ColorAttr{colUndefined, AttrUndefined},
@@ -497,6 +498,8 @@ func EmptyTheme() *ColorTheme {
 		PreviewBg:    ColorAttr{colUndefined, AttrUndefined},
 		Gutter:       ColorAttr{colUndefined, AttrUndefined},
 		PreviewLabel: ColorAttr{colUndefined, AttrUndefined},
+		Separator:    ColorAttr{colUndefined, AttrUndefined},
+		Scrollbar:    ColorAttr{colUndefined, AttrUndefined},
 	}
 }
 
@@ -516,7 +519,6 @@ func NoColorTheme() *ColorTheme {
 		Cursor:       ColorAttr{colDefault, AttrRegular},
 		Selected:     ColorAttr{colDefault, AttrRegular},
 		Header:       ColorAttr{colDefault, AttrRegular},
-		Separator:    ColorAttr{colDefault, AttrRegular},
 		Border:       ColorAttr{colDefault, AttrRegular},
 		BorderLabel:  ColorAttr{colDefault, AttrRegular},
 		Disabled:     ColorAttr{colDefault, AttrRegular},
@@ -524,6 +526,8 @@ func NoColorTheme() *ColorTheme {
 		PreviewBg:    ColorAttr{colDefault, AttrRegular},
 		Gutter:       ColorAttr{colDefault, AttrRegular},
 		PreviewLabel: ColorAttr{colDefault, AttrRegular},
+		Separator:    ColorAttr{colDefault, AttrRegular},
+		Scrollbar:    ColorAttr{colDefault, AttrRegular},
 	}
 }
 
@@ -548,7 +552,6 @@ func init() {
 		Cursor:       ColorAttr{colRed, AttrUndefined},
 		Selected:     ColorAttr{colMagenta, AttrUndefined},
 		Header:       ColorAttr{colCyan, AttrUndefined},
-		Separator:    ColorAttr{colBlack, AttrUndefined},
 		Border:       ColorAttr{colBlack, AttrUndefined},
 		BorderLabel:  ColorAttr{colWhite, AttrUndefined},
 		Disabled:     ColorAttr{colUndefined, AttrUndefined},
@@ -556,6 +559,8 @@ func init() {
 		PreviewBg:    ColorAttr{colUndefined, AttrUndefined},
 		Gutter:       ColorAttr{colUndefined, AttrUndefined},
 		PreviewLabel: ColorAttr{colUndefined, AttrUndefined},
+		Separator:    ColorAttr{colUndefined, AttrUndefined},
+		Scrollbar:    ColorAttr{colUndefined, AttrUndefined},
 	}
 	Dark256 = &ColorTheme{
 		Colored:      true,
@@ -572,7 +577,6 @@ func init() {
 		Cursor:       ColorAttr{161, AttrUndefined},
 		Selected:     ColorAttr{168, AttrUndefined},
 		Header:       ColorAttr{109, AttrUndefined},
-		Separator:    ColorAttr{59, AttrUndefined},
 		Border:       ColorAttr{59, AttrUndefined},
 		BorderLabel:  ColorAttr{145, AttrUndefined},
 		Disabled:     ColorAttr{colUndefined, AttrUndefined},
@@ -580,6 +584,8 @@ func init() {
 		PreviewBg:    ColorAttr{colUndefined, AttrUndefined},
 		Gutter:       ColorAttr{colUndefined, AttrUndefined},
 		PreviewLabel: ColorAttr{colUndefined, AttrUndefined},
+		Separator:    ColorAttr{colUndefined, AttrUndefined},
+		Scrollbar:    ColorAttr{colUndefined, AttrUndefined},
 	}
 	Light256 = &ColorTheme{
 		Colored:      true,
@@ -596,7 +602,6 @@ func init() {
 		Cursor:       ColorAttr{161, AttrUndefined},
 		Selected:     ColorAttr{168, AttrUndefined},
 		Header:       ColorAttr{31, AttrUndefined},
-		Separator:    ColorAttr{145, AttrUndefined},
 		Border:       ColorAttr{145, AttrUndefined},
 		BorderLabel:  ColorAttr{59, AttrUndefined},
 		Disabled:     ColorAttr{colUndefined, AttrUndefined},
@@ -604,6 +609,8 @@ func init() {
 		PreviewBg:    ColorAttr{colUndefined, AttrUndefined},
 		Gutter:       ColorAttr{colUndefined, AttrUndefined},
 		PreviewLabel: ColorAttr{colUndefined, AttrUndefined},
+		Separator:    ColorAttr{colUndefined, AttrUndefined},
+		Scrollbar:    ColorAttr{colUndefined, AttrUndefined},
 	}
 }
 
@@ -635,7 +642,6 @@ func initTheme(theme *ColorTheme, baseTheme *ColorTheme, forceBlack bool) {
 	theme.Cursor = o(baseTheme.Cursor, theme.Cursor)
 	theme.Selected = o(baseTheme.Selected, theme.Selected)
 	theme.Header = o(baseTheme.Header, theme.Header)
-	theme.Separator = o(baseTheme.Separator, theme.Separator)
 	theme.Border = o(baseTheme.Border, theme.Border)
 	theme.BorderLabel = o(baseTheme.BorderLabel, theme.BorderLabel)
 
@@ -645,6 +651,8 @@ func initTheme(theme *ColorTheme, baseTheme *ColorTheme, forceBlack bool) {
 	theme.PreviewFg = o(theme.Fg, theme.PreviewFg)
 	theme.PreviewBg = o(theme.Bg, theme.PreviewBg)
 	theme.PreviewLabel = o(theme.BorderLabel, theme.PreviewLabel)
+	theme.Separator = o(theme.Border, theme.Separator)
+	theme.Scrollbar = o(theme.Border, theme.Scrollbar)
 
 	initPalette(theme)
 }
@@ -677,6 +685,7 @@ func initPalette(theme *ColorTheme) {
 	ColInfo = pair(theme.Info, theme.Bg)
 	ColHeader = pair(theme.Header, theme.Bg)
 	ColSeparator = pair(theme.Separator, theme.Bg)
+	ColScrollbar = pair(theme.Scrollbar, theme.Bg)
 	ColBorder = pair(theme.Border, theme.Bg)
 	ColBorderLabel = pair(theme.BorderLabel, theme.Bg)
 	ColPreviewLabel = pair(theme.PreviewLabel, theme.PreviewBg)